move cache episodes mutation
